feat(components): :sparkles: add stepper
