+++ components/index.ts
@@ -1,1 +1,2 @@
 export * from './OnboardingLayout'
+export * from "./Stepper";
